Restructure parser
