--- ins.sh
@@ -1,2 +1,0 @@
-pip uninstall simc -y
-python setup.py install
